moved html template to server-page component
